refactor bin entrypoint, add stackblitz support
