+++ .changeset/swift-comics-hammer.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+Refactor the CLI entrypoint to support stackblitz and improve the runtime check
