+++ packages/astro/package.json
@@ -11,7 +11,7 @@
     "directory": "packages/astro"
   },
   "exports": {
-    ".": "./astro.cjs",
+    ".": "./astro.js",
     "./package.json": "./package.json",
     "./snowpack-plugin": "./snowpack-plugin.cjs",
     "./snowpack-plugin-jsx": "./snowpack-plugin-jsx.cjs",
@@ -26,7 +26,7 @@
     "#astro/*": "./dist/*.js"
   },
   "bin": {
-    "astro": "astro.cjs"
+    "astro": "astro.js"
   },
   "files": [
     "components",
@@ -60,8 +60,8 @@
     "acorn": "^7.4.0",
     "astring": "^1.7.4",
     "autoprefixer": "^10.2.5",
-    "camel-case": "^4.1.2",
     "babel-plugin-module-resolver": "^4.1.0",
+    "camel-case": "^4.1.2",
     "cheerio": "^1.0.0-rc.6",
     "ci-info": "^3.2.0",
     "del": "^6.0.0",
@@ -93,6 +93,7 @@
     "slash": "^4.0.0",
     "snowpack": "^3.8.3",
     "string-width": "^5.0.0",
+    "supports-esm": "^1.0.0",
     "tiny-glob": "^0.2.8",
     "unified": "^9.2.1",
     "yargs-parser": "^20.2.7"
